# openapi.yaml
openapi: 3.0.3
info:
  title: SiteBuilder CMS API
  description: REST API для управления сайтами, страницами, постами и медиафайлами
  version: 1.0.0
  contact:
    name: SiteBuilder Support
    url: https://github.com/sitebuilder/sitebuilder
servers:
  - url: http://localhost:3000/api
    description: Локальный сервер разработки
  - url: https://your-domain.com/api
    description: Production сервер
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
...
    Media:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        fileName:
          type: string
        url:
          type: string
        type:
          type: string
          enum: [IMAGE, VIDEO, DOCUMENT]
        mimeType:
          type: string
        size:
          type: integer
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        alt:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        userId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
paths:
...
  /media/{mediaId}:
    get:
      tags:
        - Media
      summary: Получить медиафайл по ID
      security:
        - cookieAuth: []
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Данные медиафайла
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        '401':
          description: Не авторизован
        '403':
          description: Недостаточно прав
        '404':
          description: Медиафайл не найден
    
    put:
      tags:
        - Media
      summary: Обновить медиафайл
      security:
        - cookieAuth: []
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                alt:
                  type: string
                caption:
                  type: string
      responses:
        '200':
          description: Медиафайл успешно обновлен
        '401':
          description: Не авторизован
        '403':
          description: Недостаточно прав
        '404':
          description: Медиафайл не найден
    
    delete:
      tags:
        - Media
      summary: Удалить медиафайл
      security:
        - cookieAuth: []
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Медиафайл успешно удален
        '401':
          description: Не авторизован
        '403':
          description: Недостаточно прав
        '404':
          description: Медиафайл не найден
...